requirements.txt
Django==3.2
django-filter==21.1
Pillow==8.2.0